Moved forgotten method "getCloudServices" to "GeneralCloudServiceProvider"
